replaced current encoder concept by transducer
