+++ examples/modular.yaml
@@ -28,14 +28,14 @@ modular:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 64
-      encoder: !ModularEncoder
+      encoder: !ModularSeqTransducer
         modules:
-        - !LSTMEncoder
+        - !LSTMSeqTransducer
             input_dim: 64
             hidden_dim: 64
             layers: 1
             dropout: 0.1
-        - !PyramidalLSTMEncoder
+        - !PyramidalLSTMSeqTransducer
             input_dim: 64
             hidden_dim: 64
             layers: 1
