+++ examples/pretrained.yaml
@@ -25,7 +25,7 @@ defaults:
       src_embedder: !PretrainedSimpleWordEmbedder
         filename: ../wiki.ja.vec
         emb_dim: 300
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 1
       attender: !StandardAttender
         hidden_dim: 512
