+++ examples/report.yaml
@@ -25,7 +25,7 @@ defaults:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 256
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 1
         input_dim: 256
       attender: !StandardAttender
