+++ examples/standard.yaml
@@ -27,7 +27,7 @@ defaults:
       trg_reader: !PlainTextReader {}
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder {}
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 1
       attender: !StandardAttender {}
       trg_embedder: !SimpleWordEmbedder {}
