+++ test/config/load_model.yaml
@@ -25,7 +25,7 @@ defaults:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 64
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 2
         input_dim: 64
         bidirectional: True
