+++ test/config/retrieval.yaml
@@ -26,12 +26,12 @@ defaults:
       src_embedder: !SimpleWordEmbedder
         vocab_size: 20 # TODO: set this automatically
         emb_dim: 16
-      src_encoder: !LSTMEncoder
+      src_encoder: !LSTMSeqTransducer
         bidirectional: True
       trg_embedder: !SimpleWordEmbedder
         vocab_size: 20 # TODO: set this automatically
         emb_dim: 16
-      trg_encoder: !LSTMEncoder
+      trg_encoder: !LSTMSeqTransducer
         bidirectional: True
       database: !StandardRetrievalDatabase
         reader: !PlainTextReader {}
