+++ test/config/segmenting.yaml
@@ -23,14 +23,14 @@ defaults:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 16
-      encoder: !SegmentingEncoder
-        embed_encoder: !LSTMEncoder
+      encoder: !SegmentingSeqTransducer
+        embed_encoder: !LSTMSeqTransducer
           input_dim: 16
           hidden_dim: 16
           layers: 1
           bidirectional: True
         segment_transducer: !SegmentTransducer
-          encoder: !LSTMEncoder
+          encoder: !LSTMSeqTransducer
             input_dim: 16
             hidden_dim: 16
             bidirectional: True
