+++ test/config/translator_loss.yaml
@@ -39,7 +39,7 @@ reinforce:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 32
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 1
         input_dim: 32
       attender: !StandardAttender
