+++ test/config/translator_report.yaml
@@ -26,7 +26,7 @@ defaults:
     model: !DefaultTranslator
       src_embedder: !SimpleWordEmbedder
         emb_dim: 32
-      encoder: !LSTMEncoder
+      encoder: !LSTMSeqTransducer
         layers: 1
         input_dim: 32
       attender: !StandardAttender
