+++ test/test_beam_search.py
@@ -5,7 +5,7 @@ import dynet as dy
 
 from xnmt.translator import DefaultTranslator
 from xnmt.embedder import SimpleWordEmbedder
-from xnmt.encoder import LSTMEncoder
+from xnmt.lstm import LSTMSeqTransducer
 from xnmt.attender import StandardAttender
 from xnmt.decoder import MlpSoftmaxDecoder, CopyBridge
 from xnmt.training_corpus import BilingualTrainingCorpus
@@ -27,7 +27,7 @@ class TestForcedDecodingOutputs(unittest.TestCase):
     self.model_context.dynet_param_collection = PersistentParamCollection("some_file", 1)
     self.model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context),
+              encoder=LSTMSeqTransducer(self.model_context),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100),
@@ -62,7 +62,7 @@ class TestForcedDecodingLoss(unittest.TestCase):
     self.model_context.dynet_param_collection = PersistentParamCollection("some_file", 1)
     self.model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context),
+              encoder=LSTMSeqTransducer(self.model_context),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100, bridge=CopyBridge(self.model_context, dec_layers=1)),
@@ -98,7 +98,7 @@ class TestFreeDecodingLoss(unittest.TestCase):
     self.model_context.dynet_param_collection = PersistentParamCollection("some_file", 1)
     self.model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context),
+              encoder=LSTMSeqTransducer(self.model_context),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100, bridge=CopyBridge(self.model_context, dec_layers=1)),
@@ -138,7 +138,7 @@ class TestGreedyVsBeam(unittest.TestCase):
     self.model_context.dynet_param_collection = PersistentParamCollection("some_file", 1)
     self.model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context),
+              encoder=LSTMSeqTransducer(self.model_context),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100, bridge=CopyBridge(self.model_context, dec_layers=1)),
