+++ test/test_encoder.py
@@ -5,7 +5,8 @@ import dynet as dy
 
 from xnmt.translator import DefaultTranslator
 from xnmt.embedder import SimpleWordEmbedder
-from xnmt.encoder import LSTMEncoder, ResidualLSTMEncoder
+from xnmt.lstm import LSTMSeqTransducer
+from xnmt.residual import ResidualLSTMSeqTransducer
 from xnmt.attender import StandardAttender
 from xnmt.decoder import MlpSoftmaxDecoder
 from xnmt.training_corpus import BilingualTrainingCorpus
@@ -30,13 +31,13 @@ class TestEncoder(unittest.TestCase):
   def assert_in_out_len_equal(self, model):
     dy.renew_cg()
     embeddings = model.src_embedder.embed_sent(self.training_corpus.train_src_data[0])
-    encodings = model.encoder.transduce(embeddings)
+    encodings = model.encoder(embeddings)
     self.assertEqual(len(embeddings), len(encodings))
 
   def test_uni_lstm_encoder_len(self):
     model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context, layers=3, bidirectional=True),
+              encoder=LSTMSeqTransducer(self.model_context, layers=3, bidirectional=True),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100),
@@ -46,7 +47,7 @@ class TestEncoder(unittest.TestCase):
   def test_bi_lstm_encoder_len(self):
     model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=LSTMEncoder(self.model_context, layers=1, bidirectional=False),
+              encoder=LSTMSeqTransducer(self.model_context, layers=1, bidirectional=False),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100),
@@ -56,7 +57,7 @@ class TestEncoder(unittest.TestCase):
   def test_res_lstm_encoder_len(self):
     model = DefaultTranslator(
               src_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
-              encoder=ResidualLSTMEncoder(self.model_context, layers=3),
+              encoder=ResidualLSTMSeqTransducer(self.model_context, layers=3),
               attender=StandardAttender(self.model_context),
               trg_embedder=SimpleWordEmbedder(self.model_context, vocab_size=100),
               decoder=MlpSoftmaxDecoder(self.model_context, vocab_size=100),
