+++ xnmt/segment_transducer.py
